+++ dbus-serialbattery/utils.py
@@ -14,7 +14,7 @@ import serial
 
 
 # CONSTANTS
-DRIVER_VERSION: str = "2.1.20260109dev"
+DRIVER_VERSION: str = "2.1.202601010dev"
 """
 current version of the driver
 """